requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6
gunicorn==21.2.0